refactored base code
