+++ code/.gitignore
@@ -1,0 +1,2 @@
+/out/production/CS207 Assignment 1/
+/.idea/
