+++ code/BoggleTests.java
@@ -1,17 +1,10 @@
-import java.io.ByteArrayInputStream;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.io.PrintStream;
 import java.util.*;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.lang.reflect.Field;
 
 import boggle.*;
 import boggle.Dictionary;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
+import boggle.stats.BoggleStats;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -25,7 +18,7 @@ public class BoggleTests {
         Method method = game.getClass().getDeclaredMethod("findAllWords", Map.class, Dictionary.class, BoggleGrid.class);
         method.setAccessible(true);
 
-        Dictionary boggleDict = new Dictionary("wordlist.txt");
+        Dictionary boggleDict = new Dictionary("dictionaries/wordlist.txt");
         Map<String, ArrayList<Position>> allWords = new HashMap<>();
         BoggleGrid grid = new BoggleGrid(4);
         grid.initalizeBoard("RHLDNHTGIPHSNMJO");
@@ -37,7 +30,7 @@ public class BoggleTests {
     //Dictionary Test
     @Test
     void containsWord() {
-        Dictionary dict = new Dictionary("./wordlist.txt");
+        Dictionary dict = new Dictionary("dictionaries/wordlist.txt");
         assertTrue(dict.containsWord("EnZyMe"));
         assertTrue(dict.isPrefix(""));
